+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add Makefile-extras for managing run hooks
+- Add enableSecurityInstructions option which defaults to true
 
 ### Changed
 - Upgrade Suntory to 1.3.0