+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Decouple client into message client and moderation client
+- Upgrade Suntory to 1.6.0
 
 ## 1.3.0 - 2026-06-30
 ### Added